LayoutTests/imported/w3c/web-platform-tests/user-timing/buffered-flag.any-expected.txt

CONSOLE MESSAGE: TypeError: Member PerformanceObserverInit.entryTypes is required and must be an instance of sequence
CONSOLE MESSAGE: TypeError: Member PerformanceObserverInit.entryTypes is required and must be an instance of sequence

Harness Error (FAIL), message = TypeError: Member PerformanceObserverInit.entryTypes is required and must be an instance of sequence

TIMEOUT PerformanceObserver with buffered flag sees previous marks Test timed out
TIMEOUT PerformanceObserver with buffered flag sees previous measures Test timed out

Harness Error (FAIL), message = TypeError: Member PerformanceObserverInit.entryTypes is required and must be an instance of sequence

TIMEOUT PerformanceObserver with buffered flag sees previous marks Test timed out
TIMEOUT PerformanceObserver with buffered flag sees previous measures Test timed out
